requirements.txt
fastapi
uvicorn
SQLAlchemy>=2.0.0
mysql-connector-python
python-dotenv
